+++ bouncr/db/migrate/20220228161637_create_invites.rb
@@ -7,7 +7,7 @@ class CreateInvites < ActiveRecord::Migration[6.1]
       t.boolean :inviteStatus
       t.boolean :checkinStatus
       t.integer :phoneNumber
-      t.integer :coverChargePaid
+      t.float :coverChargePaid
 
       t.timestamps
     end
